fix(ui): use mutation observer to sync height
We have been using ResizeObserver to do this, but listening on the
iFrame itself, which is not the element we should be tracking, but the
actual element is not in the DOM at the moment the listener is
registered. Using it on a single place to mitigate any performance
impact.